+++ src/request.js
@@ -1,3 +1,4 @@
+import _ from 'underscore';
 import util from 'util';
 import { checkParamOrThrow } from 'apify-client/build/utils';
 import { normalizeUrl } from 'apify-shared/utilities';
@@ -71,6 +72,13 @@ export const computeUniqueKey = (url, keepUrlFragment) => normalizeUrl(url, keep
  *   Request ID
  * @property {String} url
  *   URL of the web page to crawl.
+ * @property {String} loadedUrl
+ *   An actually loaded URL after redirects, if present. HTTP redirects are guaranteed
+ *   to be included.
+ *
+ *   When using {@link PuppeteerCrawler}, meta tag and JavaScript redirects may,
+ *   or may not be included, depending on their nature. This generally means that redirects,
+ *   which happen immediately will most likely be included, but delayed redirects will not.
  * @property {String} uniqueKey
  *   A unique key identifying the request.
  *   Two requests with the same `uniqueKey` are considered as pointing to the same URL.
@@ -99,6 +107,7 @@ class Request {
         const {
             id,
             url,
+            loadedUrl = null,
             uniqueKey,
             method = 'GET',
             payload = null,
@@ -114,6 +123,7 @@ class Request {
 
         checkParamOrThrow(id, 'id', 'Maybe String');
         checkParamOrThrow(url, 'url', 'String');
+        checkParamOrThrow(loadedUrl, 'url', 'Maybe String');
         checkParamOrThrow(uniqueKey, 'uniqueKey', 'Maybe String');
         checkParamOrThrow(method, 'method', 'String');
         checkParamOrThrow(payload, 'payload', 'Maybe Buffer | String');
@@ -130,6 +140,7 @@ class Request {
 
         this.id = id;
         this.url = url;
+        this.loadedUrl = loadedUrl;
         // NOTE: If URL is invalid, computeUniqueKey() returns null which was causing weird errors
         this.uniqueKey = uniqueKey || computeUniqueKey(url, keepUrlFragment) || url;
         this.method = method;
@@ -139,7 +150,12 @@ class Request {
         this.errorMessages = errorMessages;
         this.headers = headers;
         this.userData = userData;
-        this.handledAt = handledAt && new Date(handledAt);
+
+        // TODO: What is this text parsing good for? There's no unit test for it...
+        // eslint-disable-next-line no-nested-ternary
+        this.handledAt = _.isDate(handledAt)
+            ? handledAt
+            : (handledAt ? new Date(handledAt) : null);
     }
 
     /**