+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 xxx
 ===================
+- Added `loadedUrl` property to `Request` that contains the `request.url` after redirects.
+- Reduced default `handlePageTimeoutSecs` for both `CheerioCrawler` and `PuppeteerCrawler` to `60` seconds.
 - Added memory overload warning log message.
 - Added `keyValueStore.getPublicUrl` function.